requirements.txt
# count
numpy
torch>=1.7
opencv-python
loguru
scikit-image
tqdm
torchvision>=0.10.0
Pillow
thop
ninja
tabulate
tensorboard
lap
motmetrics
filterpy
h5py
PyMySQL
PyYaml
pycocotools
cython
pymysql-pooling
aiohttp
# cython_bbox # windows may be error

# verified versions
# onnx==1.8.1
# onnxruntime==1.8.0
# onnx-simplifier==0.3.5

# face recognition
matplotlib
easydict
opencv-contrib-python
scikit-learn
scipy
seaborn
xmltodict
basetrainer
pybaseutils
toolz

